Add `RedisCommands#ping`
