+++ src/util/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod redis;
 pub mod regex;
 pub mod serde;
